requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
GeoAlchemy2==0.14.2
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
Pillow==10.1.0
python-multipart==0.0.6
Werkzeug==3.0.1
bcrypt==4.1.2
email-validator==2.1.0
marshmallow==3.20.1
gunicorn==21.2.0
eventlet==0.33.3 